+++ src/main/java/za/co/hpsc/web/constants/SystemConstants.java
@@ -23,4 +23,6 @@ public class SystemConstants {
     public static final String ISO_DATE_TIME_FORMAT = ISO_DATE_FORMAT + " HH:mm";
     public static final String LONG_DATE_FORMAT = "dd MMMM yyyy";
     public static final String LONG_DATE_TIME_FORMAT = LONG_DATE_FORMAT + " HH:mm";
+
+    public static final String ARRAY_SEPARATOR = ";";
 }